Pass layers through props and remove LayersShowContainer
