+++ src/components/pages/settings/Settings.js
@@ -6,7 +6,7 @@ import Article from '../../layouts/Article';
 import TeamsShow from '../../teams-show/TeamsShowContainer';
 import TeamsForm from '../../teams-manager/TeamsFormContainer';
 import LayersManager from '../../layers-manager/LayersManager';
-import LayersShow from '../../layers-show/LayersShowContainer';
+import LayersShow from '../../layers-show/LayersShow';
 import Loader from '../../ui/Loader';
 
 class Settings extends React.Component {
@@ -25,7 +25,7 @@ class Settings extends React.Component {
   }
 
   render() {
-    const { team, editing, loading, saving, isManager } = this.props;
+    const { team, editing, loading, saving, isManager, publicLayers, teamLayers, userLayers } = this.props;
     return (
       <div>
         {isManager && !editing? 
@@ -43,13 +43,23 @@ class Settings extends React.Component {
                 <div className="settings-show">
                   <Article>
                     <TeamsShow />
-                    <LayersShow isManager={isManager}/>
+                    <LayersShow 
+                      isManager={isManager}
+                      publicLayers={publicLayers}
+                      teamLayers={teamLayers}
+                      userLayers={userLayers}
+                      />
                   </Article>
                 </div>
                 : 
                 <div className="settings-edit">
                   <TeamsForm team={team}/>
-                  <LayersManager isManager={isManager}/>
+                  <LayersManager 
+                    isManager={isManager}
+                    publicLayers={publicLayers}
+                    teamLayers={teamLayers}
+                    userLayers={userLayers}
+                  />
                 </div>
               }
               <Loader isLoading={saving} />
@@ -64,6 +74,9 @@ Settings.propTypes = {
   team: PropTypes.object,
   getTeam: PropTypes.func.isRequired,
   isManager: PropTypes.bool,
+  publicLayers: PropTypes.array.isRequired,
+  teamLayers: PropTypes.array.isRequired,
+  userLayers: PropTypes.array.isRequired,
   userId: PropTypes.string.isRequired
 };
